[BE] refactor(#35): 패키지 정리
