+++ backend/src/main/java/com/example/backend/auth/api/service/oauth/response/OAuthResponse.java
@@ -1,6 +1,6 @@
 package com.example.backend.auth.api.service.oauth.response;
 
-import com.example.backend.domain.mysql.define.user.constant.UserPlatformType;
+import com.example.backend.domain.define.user.constant.UserPlatformType;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
