+++ backend/src/main/java/com/example/backend/auth/config/security/filter/JwtAuthenticationFilter.java
@@ -4,8 +4,8 @@ import com.example.backend.auth.api.service.jwt.JwtService;
 import com.example.backend.common.exception.ExceptionMessage;
 import com.example.backend.common.exception.jwt.JwtException;
 import com.example.backend.common.response.JsonResult;
-import com.example.backend.domain.mysql.define.user.User;
-import com.example.backend.domain.mysql.define.user.constant.UserRole;
+import com.example.backend.domain.define.user.User;
+import com.example.backend.domain.define.user.constant.UserRole;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.SignatureException;
