+++ backend/src/main/java/com/example/backend/domain/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain.redis.config;
+package com.example.backend.domain.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
