+++ backend/src/main/java/com/example/backend/domain/define/test/repository/RedisTestRepository.java
@@ -1,6 +1,6 @@
-package com.example.backend.domain.redis.define.test.repository;
+package com.example.backend.domain.define.test.repository;
 
-import com.example.backend.domain.redis.define.test.RedisTest;
+import com.example.backend.domain.define.test.RedisTest;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RedisTestRepository extends CrudRepository<RedisTest, Long> {
