+++ backend/src/main/java/com/example/backend/domain/define/user/constant/UserRole.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain.mysql.define.user.constant;
+package com.example.backend.domain.define.user.constant;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
