--- backend/src/main/java/com/example/backend/domain/mysql/config/JpaConfig.java
@@ -1,10 +1,0 @@
-package com.example.backend.domain.mysql.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-@Configuration
-@EnableTransactionManagement
-public class JpaConfig {
-
-}
